requirements.txt
Django==2.2.3
django-filter==2.0
djangorestframework==3.10.0
djangorestframework-filters==0.11.1
drf-writable-nested==0.5.1
mysqlclient==1.4.2.post1
pytz==2019.1
sqlparse==0.3.0
